+++ src/extension.ts
@@ -1,6 +1,7 @@
 import * as vscode from 'vscode';
 import { CamelDebugAdapterDescriptorFactory } from './CamelDebugAdpaterDescriptorFactory';
 import { getRedHatService, TelemetryService } from "@redhat-developer/vscode-redhat-telemetry";
+import { CamelMavenTasksCompletionItemProvider } from './completion/CamelMavenTasksCompletionItemProvider';
 
 let telemetryService: TelemetryService;
 
@@ -8,6 +9,10 @@ const CAMEL_DEBUG_ADAPTER_ID = 'apache.camel';
 
 export async function activate(context: vscode.ExtensionContext) {
 	vscode.debug.registerDebugAdapterDescriptorFactory(CAMEL_DEBUG_ADAPTER_ID, new CamelDebugAdapterDescriptorFactory(context));
+	
+	const tasksJson:vscode.DocumentSelector = { scheme: 'file', language: 'jsonc', pattern: '**/tasks.json' };
+	vscode.languages.registerCompletionItemProvider(tasksJson, new CamelMavenTasksCompletionItemProvider());
+	
 	const redhatService = await getRedHatService(context);  
 	telemetryService = await redhatService.getTelemetryService();
 	telemetryService.sendStartupEvent();