+++ src/test/runTest.ts
@@ -11,9 +11,11 @@ async function main() {
 		// The path to test runner
 		// Passed to --extensionTestsPath
 		const extensionTestsPath = path.resolve(__dirname, './suite/index');
+		
+		const testWorkspace = path.resolve(__dirname, '../../../test Fixture with speci@l chars');
 
 		// Download VS Code, unzip it and run the integration test
-		await runTests({ extensionDevelopmentPath, extensionTestsPath });
+		await runTests({ extensionDevelopmentPath, extensionTestsPath, launchArgs: [testWorkspace, '--disable-workspace-trust'] });
 	} catch (err) {
 		console.error('Failed to run tests');
 		process.exit(1);